requirements.txt
fastapi>=0.100.0
uvicorn[standard]
pinecone>=3.0.0
sentence-transformers
python-dotenv
pydantic>=2.0.0
